+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/domain/user/dto/response/UserBasicInfoEditDto.java
@@ -1,4 +1,4 @@
-package io.f12.notionlinkedblog.domain.user.dto.info;
+package io.f12.notionlinkedblog.domain.user.dto.response;
 
 import javax.validation.constraints.NotEmpty;
 
@@ -11,7 +11,8 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class UserBlogTitleEditDto {
+public class UserBasicInfoEditDto {
 	@NotEmpty
-	private String blogTitle;
+	private String username;
+	private String introduction;
 }
