+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/domain/user/dto/response/UserSocialInfoEditDto.java
@@ -1,4 +1,4 @@
-package io.f12.notionlinkedblog.domain.user.dto.info;
+package io.f12.notionlinkedblog.domain.user.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
